mexendo no modulo disciplina

+++ Modulos/metodos.c
@@ -22,7 +22,7 @@ void mostraPessoa(Escola *escola, int contador)
 }
     int menu1 = 0;
     int matricula = 0;
-    int pis = 0;
+    long int pis = 0;
     printf("\n1-Aluno \n2-Professor \n3-Voltar \n");
     scanf("%i", &menu1);
     getchar();
@@ -68,7 +68,7 @@ void mostraPessoa(Escola *escola, int contador)
         break;
     case 2:
         printf("Digite o PIS: ");
-        scanf("%d", &pis);
+        scanf("%ld", &pis);
         getchar();
         for (int i = 0; i < contador; i++)
         {
